refactor: moves shot zone results
